pyproject.toml
[project]
name = "interlock-backend"
version = "0.1.0"
description = "Identity Provider and LDAP Manager Backend"
authors = [
    {name = "Dylan Blanque",email = "dylan.blanque@gmail.com"}
]
license = {text = "GNU AGPLv3"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "appdirs (==1.4.4)",
    "asgiref (==3.7.2)",
    "attrs (==23.1.0)",
    "autopep8 (==2.0.2)",
    "babel (==2.12.1)",
    "blinker (==1.6.2)",
    "cached-property (==1.5.2)",
    "certifi (==2023.7.22)",
    "cffi (==1.15.1)",
    "chardet (==5.2.0)",
    "charset-normalizer (==3.2.0)",
    "click (==8.1.6)",
    "colorama (==0.4.6)",
    "coreapi (==2.3.3)",
    "coreschema (==0.0.4)",
    "cryptography (==41.0.3)",
    "defusedxml (==0.7.1)",
    "django (==4.2.4)",
    "django-choices (==2.0.0)",
    "django-common-helpers (==0.9.2)",
    "django-cors-headers (==4.2.0)",
    "django-cron (==0.6.0)",
    "django-extensions (==3.2.3)",
    "django-filter (==23.2)",
    "django-oidc-provider (==0.8.3)",
    "django-otp (==1.2.2)",
    "django-phonenumber-field (==7.1.0)",
    "django-python3-ldap (==0.15.5)",
    "django-ratelimit (==4.1.0)",
    "django-reversion (==5.0.4)",
    "django-solo (==2.1.0)",
    "django-sslserver (==0.22)",
    "djangorestframework (==3.14.0)",
    "djangorestframework-simplejwt (==5.2.2)",
    "dnspython (==2.4.1)",
    "exceptiongroup (==1.1.2)",
    "flask (==2.3.2)",
    "future (==0.18.3)",
    "hypothesis (==6.82.0)",
    "idna (==3.4)",
    "impacket (==0.10.0)",
    "inflection (==0.5.1)",
    "isodate (==0.6.1)",
    "itsdangerous (==2.1.2)",
    "itypes (==1.2.0)",
    "jinja2 (==3.1.2)",
    "ldap3 (==2.9.1)",
    "ldapdomaindump (==0.9.4)",
    "lxml (==4.9.3)",
    "markdown (==3.4.4)",
    "markupsafe (==2.1.3)",
    "naked (==0.1.32)",
    "packaging (==23.1)",
    "pbr (==5.11.1)",
    "pip-autoremove (==0.10.0)",
    "platformdirs (==3.10.0)",
    "psycopg2-binary (==2.9.6)",
    "py-bcrypt (==0.4)",
    "py-moneyed (==3.0)",
    "pyasn1 (==0.4.8)",
    "pycodestyle (==2.11.0)",
    "pycparser (==2.21)",
    "pycryptodome (==3.21.0)",
    "pycryptodomex (==3.18.0)",
    "pyjwkest (==1.4.2)",
    "pyjwt (==2.8.0)",
    "pyopenssl (==23.2.0)",
    "pyparsing (==3.1.1)",
    "pytz (==2023.3)",
    "pyyaml (==6.0.1)",
    "requests (==2.31.0)",
    "requests-file (==1.5.1)",
    "requests-toolbelt (==1.0.0)",
    "ruamel-yaml (==0.17.32)",
    "ruamel-yaml-clib (==0.2.7)",
    "shellescape (==3.8.1)",
    "six (==1.16.0)",
    "sortedcontainers (==2.4.0)",
    "sqlparse (==0.4.4)",
    "termcolor (==2.3.0)",
    "toml (==0.10.2)",
    "tomli (==2.0.1)",
    "typing-extensions (==4.7.1)",
    "uritemplate (==4.1.1)",
    "urllib3 (==2.0.4)",
    "werkzeug (==2.3.6)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "ruff (>=0.11.0,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.10.0"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "interlock_backend.settings"

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
docstring-code-format = true

[tool.coverage.run]
omit = [
    "manage.py",
    "interlock_backend/asgi.py",
    "interlock_backend/wsgi.py",
]
